hippo/requirements.txt
levenshtein
datasets
evaluate
streamlit
st-annotated-text
htbuilder
wandb
torch
torchaudio
torchvision
transformers
sacrebleu
accelerate
black
openai
orjson
peft
bitsandbytes
scipy